requirements.txt
Flask~=3.1.0
Faker~=37.1.0
Flask-SQLAlchemy~=3.1.1
psycopg2~=2.9.10
SQLAlchemy~=2.0.40
python-dotenv~=1.1.0